moved sdl_mixer into audiomodule.hpp
