+++ engine/moduleManager/modules/audio/AudioModule.hpp
@@ -5,10 +5,22 @@
 #ifndef AUDIOMODULE_HPP
 #define AUDIOMODULE_HPP
 #include <map>
+#include <SDL_mixer.h>
 #include <string>
-// #include <Audio.hpp>//todo: Audio.hpp
 #include <components/audio/Audio.hpp>
 #include <moduleManager/ModuleWrapper.hpp>
+
+struct AudioFragment {
+	AudioFragment(std::string path, uint8_t volume, bool isLooping) : path(std::move(path)), volume(volume),
+	                                                                  isLooping(isLooping) {}
+
+	std::string path;
+	uint8_t volume;
+	bool isLooping;
+	Mix_Chunk* audioChunk{};
+	int playingChannel{};
+	uint8_t numberOfInstances{};
+};
 
 class AudioModule : public ModuleWrapper {
 public:
